pyproject.toml
[tool.poetry]
name = "research_agent"
version = "0.1.0"
description = "An AI Research Assistant powered by LLMs"
authors = ["BLT <letang.boon@gmail.com>"]
readme = "README.md"
packages = [{include = "research_agent"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
pydantic = "^2.5.3"
tqdm = "^4.66.1"
litellm = "^1.59.0"
pydantic-settings = "^2.7.1"
python-dotenv = "^1.0.1"
langchain = "^0.3.15"
langchain-openai = "^0.3.2"
pandas = "^2.2.3"
matplotlib = "^3.10.0"
tabulate = "^0.9.0"
rich = "^13.9.4"
langgraph = "^0.2.67"
tweepy = {extras = ["async"], version = "^4.15.0"}

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"
pytest = "^8.0.0"
pytest-cov = "^4.1.0"
ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
lint.select = ["E", "F"]    # Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
lint.ignore = [
    "C90", # McCabe complexity
    "CPY", # Missing copyright
    "COM812", # Messes with the formatter
    "ERA", # Eradicate commented-out code
    "FIX002", # Line contains TODO
    "ISC001", # Messes with the formatter
    "PERF203", # Rarely useful
    "PLR09", # Too many something (arg, statements, etc)
    "RUF012", # Pydantic models are currently not well detected. See https://github.com/astral-sh/ruff/issues/13630
    "TD002", # Missing author in TODO
    "TD003", # Missing issue link in TODO
    "TRY301", # A bit too harsh (Abstract `raise` to an inner function)

    # Rules that are TODOs
    "ANN",
]

# Allow unused variables when underscore-prefixed.
lint.dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"